+++ ZenWare.DLL/src/Features/EnginePrediction/EnginePrediction.cpp
@@ -34,7 +34,9 @@ void CFeatures_EnginePrediction::Start(C_BasePlayer* pLocal, CUserCmd* cmd)
 
 	if (cmd->weaponselect != 0)
 	{
-		C_BaseCombatWeapon* pWeapon = I::ClientEntityList->GetClientEntity(cmd->weaponselect)->As<C_BaseCombatWeapon*>();
+		C_BaseCombatWeapon* pWeapon = nullptr;
+		if (IClientEntity* pWepEnt = I::ClientEntityList->GetClientEntity(cmd->weaponselect))
+			pWeapon = pWepEnt->As<C_BaseCombatWeapon*>();
 
 		if (pWeapon)
 			pLocal->SelectItem(pWeapon->GetName(), cmd->weaponsubtype);